requirements.txt
PyNaCl==1.5.0
discord.py==1.7.3
youtube_dl==2021.12.17
python-decouple==3.6
